Fix bugs in AddWinBordersToMonitorWorkArea
Testing demonstrated that the approach used in the function to inferring actual visible window width from client and outer rectangle coordinates was inconsistent depending on the nature of the application that was active. A workaround was devised and implemented that depends on basing border widths off of the left and bottom edges of the active window, which so far have proven consistently determined only be border width in applications tested.