Add hotkey >^!+#Up
Add hotkey for snapping the active window to the top edge of the monitor; moreover, if it is already snapped, the hotkey increases the window's height up to a certain threshold. Additionally, the hotkey ensures that the width of the window fills the full horizontal extent of its monitor's work area.