Refactor hotkey ^!#Numpad5
Redesign sequence of operations such that the activewindow is snapped to the center of its monitor; moreover, if it is already snapped, decrement its width and height up to certain minimum thresholds, and wrap the width/height of those thresholds are exceeded. Also, make decrements proportional to the width and height of the monitor's work area.