Modify usage of key parameter in CommitCssBuild function
Instead of assuming that the parameter fpGitFolder (which contains the file path to the git folder in which CSS build related files will be committed) is relative to the git folder, assume it is a complete file path. This enables different git folder root locations to be utilized with the function.