Modify method for updating the CSS development submodule
Modify the git commands utilized to update the CSS development submodule that contains CSS build dependencies used in all DAESA website development projects. Specifically, use a different combination of git commands so that the submodule can be updated without the need to commit a change that represents nothing more than copying the latest committed state of the project.