Update VIM-styled cursor movement functionality
Add a toggle state for triggering a "VIM-y" or VIM-styled cursor movement mode. When active, this mode overrides the default behavior of u, i, j, k, l, semicolon, m, and comma keys to mimic VIM's benefit of moving the cursor without needing to move the right hand to the arrow keys on a standard QWERTY keyboard. Moreover, use an implementation that employs the tilde modifier to avoid distrupting hotstrings that include these keys.