Fix more bugs caused by VIMy mode implementation
The implementation of a persistent VIMy mode, which relies on the Suspend command, was found to intefere with entry of hotstrings containing characters associated with a VIMy mode hotkey. (The only exception was a hotstring that simply ended with a VIMy mode hotkey, such as `@dogl`.) Thus, it does not look like a robust implementation of VIMy mode is possible without significantly breaking existing functionality. Consequently, switch to an implementation that relies on context-sensitivity and the ampersand hotkey modifier.