Fix bug in how system monitors are arranged by the script
Fix bugs in the functions SetMonitorBounds and SetMonitorWorkAreas, which are intended to order monitors according to their spatial configuration in reality. After adding a fourth monitor to my system, it was discovered that these functions do not consider the vertical dimension. Thus, implement a more complex sorting process where the monitors are first arranged such that monitors on top come first in the final arrangement, and then furher arranged the monitors on the same vertical row from left to right.