Modify behavior of hotkey for snapping the mouse cursor
Modify behavior of `^#LButton` and `!SC029` hotkeys for snapping the mouse cursor; instead of repositioning the cursor at the left edge of the active window, position the cursor at the midpoint of the monitor the cursor is on. Moreover, if the cursor is already at the midpoint, move it to the midpoint on the next monitor installed on the system. This change in behavior is useful for rapidly moving the mouse between a complex spatial arrangement of system monitors.