Enhancements: Make several improvements to PowerShell profile
Make the following enhancements and bug fixes to PowerShell/Microsoft.PowerShell_profile.ps1: 1) Add LastWriteTime as a basis for file comparison to the command Compare-Directories. 2) Add more git diff invocation commands, along with accompanying aliases, that generate a log file as output. 3) Tweak the messaging used in the command Write-Welcome-Msg-to-Host.