Refactor: PowerShell commands for opening DAESA websites
In PowerShell/Microsoft.PowerShell_profile.ps1, first update Get-Array-of-Daesa-Website-Urls with a reclassification of the Advising website as associated with  DAESA. Next, add parameters to Get-Array-of-GitHub-Repos that enable GitHub folders to be filtered. Finally, add a parameter to Open-Daesa-Website that allows the invoker to open websites as tabs in the last activated Chrome window, if any, instead of a new window.