Bug fix: RegEx needle used in Hotstring backupOuePage
In Coding/frontEndCoding.ahk, update a regular expression search string used in hotstring backupOuePage so that Spine theme page builder rows that use a section wrapper are captured as valid markup. Also, replace the splash text message that concludes the backup process with a GUI-based message box to reduce the likelihood that the user will miss seeing that the process is complete.