Fix: Hotstring sendActiveWinToBack
In «WorkspaceManagement/workspaceManagement.ahk», add a slight offset to the mouse positioning after the next window below the previously active window that was sent to the bottom of the stack gets activated. This prevents unintended behavior from the triggering of the #Numpad5 hotkey when windows when the «sendActiveWinToBack» is repeatedly triggered.