Feat: Web request cmdlets → PowerShell profile
In «PowerShell/Microsoft.PowerShell_profile.ps1», add two new cmdlets for facilitating web requests: «Save-Web-Page-Source-Code» and «Test-Web-Page-Uri-Format». Write «Test-Web-Page-Uri-Format» to verify that a string representing a prospective URI conforms to spec. Write «Save-Web-Page-Source-Code» to use the «Invoke-WebRequest» cmdlet to process an array of URIs to web pages such that the content of each page is stored to an HTML source code file at the current location.