Feat: Add'l text replacement hotkey
In «TextInput/textReplacement.ahk», add a hotkey for inserting an en dash using the hyphen key. Also modify the SendInput sequence for inserting a facepalm emoji so that the active font is preserved in Microsoft Office products (in cases where the Georgian letter "las" used as the hand symbol is not present in the character set of the active font, a font substitution is automatically made).